webserver/requirements.txt
pytube
openai-whisper
streamlit
docx
comtypes